# docs/swagger.yaml
definitions:
...
  models.VmsMasVehicleIsActiveUpdate:
    properties:
      is_active:
        example: "1"
        type: string
      mas_vehicle_uid:
        example: f3b29096-140e-49dc-97ee-17fa9352aff6
        type: string
    type: object
...
info:
  contact: {}
  title: VMS_PLUS
  version: "1.0"
paths:
...
  /api/vehicle-info/{mas_vehicle_uid}:
    get:
      consumes:
      - application/json
      description: Get vehicle info
      parameters:
      - description: MasVehicleUID (mas_vehicle_uid)
        in: path
        name: mas_vehicle_uid
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Get vehicle info
      tags:
      - Vehicle
...
  /api/vehicle-management/search:
    get:
      consumes:
      - application/json
      description: Get a list of vehicles filtered by license plate, brand, or model
        with pagination
      parameters:
      - description: vehicle_license_plate,vehicle_brand_name,vehicle_model_name to
          search
        in: query
        name: search
        type: string
      - description: Filter by vehicle owner department SAP
        in: query
        name: vehicle_owner_dept_sap
        type: string
      - description: Filter by Car type
        in: query
        name: ref_vehicle_category_code
        type: string
      - description: Filter by vehicle status code (comma-separated, e.g., '1,2')
        in: query
        name: ref_vehicle_status_code
        type: string
      - description: Filter by ref_fuel_type_id
        in: query
        name: ref_fuel_type_id
        type: string
      - description: Order by vehicle_license_plate, vehicle_mileage, age,is_active
        in: query
        name: order_by
        type: string
      - description: 'Order direction: asc or desc'
        in: query
        name: order_dir
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of records per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Get vehicles by license plate, brand, or model with pagination
      tags:
      - Vehicle-management
  /api/vehicle-management/update-vehicle-is-active:
    put:
      consumes:
      - application/json
      description: This endpoint updates the active status of a vehicle.
      parameters:
      - description: VmsMasVehicleIsActiveUpdate data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VmsMasVehicleIsActiveUpdate'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Update vehicle active status
      tags:
      - Vehicle-management
  /api/vehicle/{mas_vehicle_uid}:
    get:
      consumes:
      - application/json
      description: This endpoint allows a user to retrieve the details of a specific
        vehicle associated with a booking request.
      parameters:
      - description: MasVehicleUID (mas_vehicle_uid)
        in: path
        name: mas_vehicle_uid
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Retrieve details of a specific vehicle
      tags:
      - Vehicle
  /api/vehicle/departments:
    get:
      consumes:
      - application/json
      description: Fetches a list of departments grouped by dept_sap, including dept_short
        and dept_full
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Get department list
      tags:
      - Vehicle
  /api/vehicle/search:
    get:
      consumes:
      - application/json
      description: Retrieves vehicles based on search text, department, and car type
        filters
      parameters:
      - description: Search text (Vehicle Brand Name or License Plate)
        in: query
        name: search
        type: string
      - description: Filter by icle Owner Department
        in: query
        name: vehicle_owner_dept
        type: string
      - description: Filter by Car Type
        in: query
        name: car_type
        type: string
      - description: Filter by Vehicle Category Code
        in: query
        name: category_code
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of records per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Search vehicles by brand, license plate, and filters
      tags:
      - Vehicle
  /api/vehicle/types:
    get:
      consumes:
      - application/json
      description: Fetches all vehicle types, optionally filtered by name
      parameters:
      - description: Filter by vehicle type name (partial match)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Get vehicle types
      tags:
      - Vehicle
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-ApiKey
    type: apiKey
  AuthorizationAuth:
    description: Bearer [your Authorization]
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"